+++ vhost-user-backend/CHANGELOG.md
@@ -4,6 +4,8 @@
 
 ### Added
 ### Changed
+- [[316]](https://github.com/rust-vmm/vhost/pull/316) Use mio to replace Epoll. Expose event_loop::EventSet.
+
 ### Deprecated
 ### Fixed